moved utils directory to package top level
