+++ readnext/data/s2_preprocess_documents_chunks.py
@@ -10,7 +10,7 @@ import pandas as pd
 
 from readnext.config import DataPaths
 from readnext.data.preprocessing_utils import add_rank
-from readnext.modeling.utils import setup_progress_bar
+from readnext.utils import setup_progress_bar
 
 
 def flatten_list_of_dicts(list_of_dicts: list[dict], key: str) -> list[str]:
