+++ readnext/data/s4_combine_documents_labels_chunks.py
@@ -3,7 +3,7 @@
 import pandas as pd
 
 from readnext.config import DataPaths
-from readnext.modeling.utils import setup_progress_bar
+from readnext.utils import setup_progress_bar
 
 
 def main() -> None:
