+++ readnext/data/s8_combine_documents_authors_labels_citations_chunks.py
@@ -6,7 +6,7 @@ dataframe.
 import pandas as pd
 
 from readnext.config import DataPaths
-from readnext.modeling.utils import setup_progress_bar
+from readnext.utils import setup_progress_bar
 
 
 def main() -> None:
