+++ readnext/modeling/language_models/s2_run_embedder.py
@@ -6,6 +6,7 @@ from sklearn.feature_extraction.text import TfidfVectorizer
 from transformers import BertModel
 
 from readnext.config import ModelPaths, ModelVersions, ResultsPaths
+from readnext.modeling import save_df_to_pickle
 from readnext.modeling.language_models import (
     BERTEmbedder,
     BERTTokenizer,
@@ -16,7 +17,6 @@ from readnext.modeling.language_models import (
     Word2VecEmbedder,
     embeddings_mapping_to_frame,
 )
-from readnext.modeling.utils import save_df_to_pickle
 
 
 def main() -> None:
